fix: 修复 clippy 警告和格式问题
- 修复 PI 常量近似值警告
- 修复相同 if 块警告
- 修复循环变量只用于索引的警告
- 修复不必要的类型转换
- 改进代码格式

+++ services/comsrv/src/service/reconnect.rs
@@ -180,7 +180,7 @@ impl ReconnectHelper {
 
     /// 计算下次重试延迟
     pub fn calculate_next_delay(&self) -> Duration {
-        let attempt = self.state.current_attempt.saturating_sub(1) as u32;
+        let attempt = self.state.current_attempt.saturating_sub(1);
         let base_delay = self.policy.initial_delay;
         let multiplier = self.policy.backoff_multiplier;
